Show and delete uploaded deliveries for user
