+++ src/Geopilot.Frontend/public/locale/de/common.json
@@ -87,6 +87,7 @@
   "maxFileCount_other": "bis zu {{count}} Dateien",
   "maxPerFile": "{{size}} pro Datei",
   "maxTotalSize": "{{size}} insgesamt",
+  "myDeliveries": "Meine Lieferungen",
   "name": "Name",
   "noMandatesFound": "Es wurde kein passendes Mandat für Ihre Lieferung gefunden.",
   "or": "oder",
