+++ src/Geopilot.Frontend/src/pages/user/deliveries/userDeliveryOverview.tsx
@@ -1,0 +1,20 @@
+import { Typography } from "@mui/material";
+import { useTranslation } from "react-i18next";
+import { DeliveryGrid } from "../../../components/grids/deliveryGrid.tsx";
+import { CenteredBox } from "../../../components/styledComponents.ts";
+import { styled } from "@mui/system";
+
+const DeliveryOverviewBox = styled(CenteredBox)({
+  height: "100%",
+});
+
+export const UserDeliveryOverview = () => {
+  const { t } = useTranslation();
+
+  return (
+    <DeliveryOverviewBox>
+      <Typography variant="h2">{t("myDeliveries")}</Typography>
+      <DeliveryGrid fetchUrl="/api/v1/delivery/uploads" columns={["date", "mandateName", "comment"]} />
+    </DeliveryOverviewBox>
+  );
+};
